micro-svc-backend/requirements.txt
mysql-connector-python
Flask==2.3.2
boto3==1.26.153




